update db price column type, fix bugs

+++ ShopWebsite.DAL/Models/ProductModels/Product.cs
@@ -15,7 +15,7 @@ namespace ShopWebsite.DAL.Models.ProductModels
         public string Alias { get; set; }
         public string Detail { get; set; }
         public int ManufactureYear { get; set; }
-        public long Price { get; set; }
+        public double Price { get; set; }
         public long View { get; set; } = 0;
         public long PurchaseCounter { get; set; } = 0;
         public ProductType Type { get; set; }
